fixed: when updating the product, if you have not pinned the file with the overall dimensions again, it is deleted.

+++ service/AdminService.js
@@ -189,7 +189,7 @@ class AdminService{
                 await sizeService.create(processedImages[0].path, product.id)
             }
         }
-        else{
+        else if(!product.size?.id){
             const old = await sizeService.get(product.id)
             if(old){
                 await sizeService.deleteFile(product.id)
